Rename `yokogawa/metadata_parsing.py` into `yokogawa/metadata.py`
