+++ examples/tools/show_FOV_ROIs/show_FOV_ROIs.py
@@ -4,7 +4,7 @@ An example of visualizing FOV ROIs and their overlaps.
 import matplotlib.pyplot as plt
 
 from fractal_tasks_core.lib_ROI_overlaps import run_overlap_check
-from fractal_tasks_core.yokogawa.metadata_parsing import (
+from fractal_tasks_core.yokogawa.metadata import (
     parse_yokogawa_metadata,
 )
